feat: migrate group media ops to session-owned MediaOps (Phase 11)
