+++ src/integration_tests/test_cases/chat_media_upload/receive_message_with_media.rs
@@ -56,6 +56,7 @@ impl ReceiveMessageWithMediaTestCase {
 
 #[async_trait]
 impl TestCase for ReceiveMessageWithMediaTestCase {
+    #[allow(deprecated)]
     async fn run(&self, context: &mut ScenarioContext) -> Result<(), WhitenoiseError> {
         tracing::info!(
             "Testing receive message with media: sender={}, receiver={}, group={}",
